Do not follow symlinks.
In this approach, a symlink is uploaded to the server as a regular file that contains the link target.  Since it is with mode 777, we treat all files with such a mode as a link.  This is not completely satisfactory, since a nonmodified client will change the mode and overwrite the file.